app/CRM/CRM-server/pyproject.toml
[project]
name = "crm-server"
version = "0.1.0"
description = "CRM-server — Customer relationship management with A2A agents for profiling, recommendation, order, and AI management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "a2a-sdk[http-server]>=0.2.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
